refactor: Organize classes into appropriate packages
