+++ src/main/java/ru/clevertec/command/PreparePizzaCommand.java
@@ -1,7 +1,7 @@
 package ru.clevertec.command;
 
 import lombok.AllArgsConstructor;
-import ru.clevertec.PizzaPreparationService;
+import ru.clevertec.service.PizzaPreparationService;
 import ru.clevertec.model.Pizza;
 import ru.clevertec.model.PizzaOrderRequest;
