+++ src/main/java/ru/clevertec/config/PizzeriaConfig.java
@@ -1,4 +1,4 @@
-package ru.clevertec;
+package ru.clevertec.config;
 
 public class PizzeriaConfig {
     public static final int MAX_WORKERS = 2;
