+++ src/main/java/ru/clevertec/order/OrderManager.java
@@ -1,4 +1,4 @@
-package ru.clevertec;
+package ru.clevertec.order;
 
 public interface OrderManager {
     void takeOrder(String type, String clientName, boolean withCheeseBoard, boolean withDoubleCheese);
