+++ src/main/java/ru/clevertec/order/PizzaOrderManager.java
@@ -1,7 +1,10 @@
-package ru.clevertec;
+package ru.clevertec.order;
 
+import ru.clevertec.service.PizzaPreparationService;
+import ru.clevertec.service.PizzaPreparationServiceImpl;
 import ru.clevertec.command.Command;
 import ru.clevertec.command.PreparePizzaCommand;
+import ru.clevertec.config.PizzeriaConfig;
 import ru.clevertec.decorator.CheeseBoardDecorator;
 import ru.clevertec.decorator.DoubleCheeseDecorator;
 import ru.clevertec.model.PizzaOrderRequest;
